refactor(ui): two-level lazy split for MFE icon chunks
Move import.meta.glob out of lazyIconRegistry.tsx (which is in the main
bundle) into a new iconLoader.ts that is itself lazily imported. The glob
map of 1500+ filenames now lives in the iconLoader chunk, not in the main
bundle. Per-icon stubs in lazy-icons/ are still individual async chunks,
so users download only the specific icons an admin configured.

Load path: main bundle → iconLoader.ts (glob map) → lazy-icons/<Name>.ts
